Fixed some warnings both on server and on client side

+++ Server/SpawnComponent.cpp
@@ -20,9 +20,9 @@ SpawnComponent SpawnOneComponent(std::vector<Component>& components, std::vector
 		DirectX::XMFLOAT3 tmp(0.f, 0.0f, 0.0f);
 
 
-		tmp.x = rand() % 201 - 100;
-		tmp.y = rand() % 201 - 100;
-		tmp.z = rand() % 201 - 100;
+		tmp.x = (float)(rand() % 201 - 100);
+		tmp.y = (float)(rand() % 201 - 100);
+		tmp.z = (float)(rand() % 201 - 100);
 
 		randomPos = normalizeXMFLOAT3(tmp);
