+++ StortSpelProjekt/SpaceShip.cpp
@@ -117,7 +117,7 @@ void SpaceShip::takeOff()
 {
 	//Icon initiation
 	DirectX::XMFLOAT3  pos = this->position;
-	static float constant = this->position.Length() *1.2;
+	static float constant = (float)(this->position.Length() *1.2);
 	DirectX::XMFLOAT3 upDir = this->getUpDirection();
 	constant += 0.1f;
 	DirectX::XMFLOAT3 test(upDir.x * constant, upDir.y * constant, upDir.z * constant);
